refactor: Move spherical data contouring to stats
